recipes-core/images: build base-files ahead of the image
Use ISAR_REBUILD_ESSENTIAL_PKGS in conf/layer.conf: this does not
cause Isar/bitbake to build our base-files recipe ahead of the
image.

Then added "base-files" to DEPENDS of the isar-image-base recipe
(using a bbappend): package was built as expected and added to
isar-apt. The image manifest however shows that the version in
the image is: 12.4+deb12u2; was expecting 4:1.0 (because our version
has a greater epoch setting).

Signed-off-by: Cedric Hombourger <cedric.hombourger@siemens.com>